Converted dashboard JSX to individual components
